Bug fix:
When Vim's search feature is relied up on to find search results
within the search results, characters that are normally treated
as special are now not. For example searching for ^ works correctly
now for both what shows up in the search results and the
highlighting withing them and using n to go to next result in the
results.