Behavioral improvement / bug fix
This is kind of a bug fix and kind of a behaviroal change. I guess
the most correct way to see I both changed / improved the behavior
and in so doing fixed a bug in the documention.

Now when you are in the search result window and you type enter,
not only are you taken to the correct file and line number and the
search term sometimes highlighted, but now the cursor also moves to
the search term on that line. The documetation has for a while now
claimed that this is how this plugin behaves. Somehow I have had it
in my head for a while now that it used to behave this way, but
from my testing, as far as I can tell, it never did. The
documetation is finally accurate because the plugin actually works
this way now.