You can edit and save search results:
The search results buffer is parsed for a file and line number. If a
matching file and line number can be found, the following contents will
be compared to what is currently stored in that line in the file system.
If they are different, the new contents as found in the search results
will replace that line. After all lines have been replaced in this way,
any open buffers to those files will be reloaded to dipslay the latest
file contents, but only if the buffer contents had not been modified
since the last save.