Prevent result buffers from staying around:
Result buffers get wiped when the buffer gets hidden, or the window is
closed. This is the only way I could find that would warn you if you try
to quit the result window with unsaved changes that would also not leave
result buffers around. The problem with leaving result buffers around is
they keep building up as you perform more and more searches (unless you
delete them), and they quickly become out of date as you edit code.
While deleting and wiping these buffers may not always be preferable, it
is most of the time, and you can learn to expect this and work it into
your workflow.