Rewrote the code to navigate to the search result, fixing multiple bugs and improving behavior:
Previously when you pressed the enter key to navigate to a search
result, some assumptions were made about your window layout, which
sometimes resulted in buggy or unuseful behavior. Now it alays looks up
the results in a window above the results window. If the window
currently directly above the results window is not also a results
window, that window will be used. Otherwise, a new window will be
created directly above the results window to display the file with the
result. If no other windows are displayed in the current tab, or if
there are no other windows at all, a new window is still opened above
the search results.